refactor(hooks/): rename handlers.js to handler.js
closes #53
